+++ adminer/include/driver.inc.php
@@ -34,7 +34,7 @@ abstract class SqlDriver {
 	/** Connect to the database
 	* @return Db|string string for error
 	*/
-	static function connect(?string $server, string $username, string $password) {
+	static function connect(string $server, string $username, string $password) {
 		$connection = new Db;
 		return ($connection->attach($server, $username, $password) ?: $connection);
 	}
@@ -275,7 +275,7 @@ AND CHECK_CLAUSE NOT LIKE '% IS NOT NULL'", $this->conn); // ignore default IS N
 	}
 
 	/** Get all fields in the current schema
-	* @return array<list<array{field:string, null:bool, type:string, length:?numeric-string, primary?:numeric-string}>>
+	* @return array<list<array{field:string, null:bool, type:string, length:?numeric-string}>> optionally also 'primary'
 	*/
 	function allFields(): array {
 		$return = array();