+++ CHANGELOG.md
@@ -1,4 +1,17 @@
 ## Adminer dev
+- PostgreSQL: Mark unique partial indexes as unique (bug #1172)
+- ClickHouse: Fix offset (bug #1188)
+- ClickHouse: Fix list of tables (bug #1176)
+- Plugins: Methods showVariables() and showStatus() (bug #1157)
+
+## Adminer 5.4.1 (released 2025-09-26)
+- SQL command: Unlink NULL primary keys
+- Do not quote 0 in CSV export
+- Warn about exceeded upload_max_filesize in imports
+- Prolong queries saved from SQL command to URL (bug #1154)
+- MySQL: Fix displaying routine definition (bug #1156, regression from 5.4.0)
+
+## Adminer 5.4.0 (released 2025-09-08)
 - Allow specifying operator in search anywhere
 - Do not order descending in GROUP BY select
 - Allow exporting SQL in SQL command (bug #1092)
@@ -7,6 +20,11 @@
 - Display @ after username without server in existing logins
 - Display data length and index length for materialized views
 - Link routines from syntax highlighting
+- Autofocus added field in alter table
+- Executed SQL commands: Add button for copy to clipboard
+- Load more: run syntax highlighter
+- Allow connecting to IPv6 (bug #1095)
+- MySQL: Fix saving empty enum (bug #1152)
 - MySQL 5.0-: Do not load partitioning info in alter table (bug #1099)
 - MariaDB: Parse COLLATE in routine definition (bug #1104)
 - PostgreSQL: Show structure of inherited tables
@@ -16,7 +34,11 @@
 - PostgreSQL: Allow comparing json columns (bug #1107)
 - PostgreSQL: Shorten values in hstore columns
 - PostgreSQL: Quote edit value with interval operator
-- PostgreSQL 11-: Avoid duplicate oid in table status (bug #1089)
+- PostgreSQL: Fix calling functions with name-less parameters
+- PostgreSQL: Fix calling functions returing table
+- PostgreSQL: Don't treat user types containing 'file' as blobs (bug #1118)
+- PostgreSQL: Export DROP and CREATE DATABASE (bug #1140)
+- PostgreSQL 11-: Avoid duplicate oid in table status (bug #1089, regression from 5.3.0)
 - Elasticsearch: Support dropping aliases
 - Plugins: Methods afterConnect(), processList() and killProcess()
 - New plugin: Display row numbers in select (bug #1106)
@@ -140,8 +162,8 @@
 - MariaDB: Fix creating and altering generated columns (bug #897)
 - PostgreSQL: Fix "where" and "order" privileges (bug #902, regression from 5.0.2)
 - SQLite: Fix creating table in compiled version (bug #901, regression from 5.0.0)
-- Elastic: Do not pass null values on insert (PR #892)
-- Elastic: Fix displaying sparse rows (PR #893)
+- Elasticsearch: Do not pass null values on insert (PR #892)
+- Elasticsearch: Fix displaying sparse rows (PR #893)
 - Plugins: Add method dumpFooter()
 
 ## Adminer 5.0.2 (released 2025-03-10)
@@ -222,6 +244,7 @@
 - SQLite: Fix expressions in default values (bug SF-860)
 - MS SQL: Foreign keys in non-default schema (bug SF-833)
 - Oracle: Include tables granted by other user
+- Elasticsearch: Move to plugin
 - MongoDB: Execute commands against the selected DB
 
 ## Adminer 4.15.0